Move TreeNode to own file
